Fix code review issues: UInt128 balances, batch count accuracy, linearizable error message
Agent-Logs-Url: https://github.com/evilz/Tigerbeetle-sample/sessions/72167f86-41b5-411e-925c-59c15d341be8

+++ src/TigerBeetleSample.Api/Endpoints/PerformanceEndpoints.cs
@@ -116,8 +116,8 @@ public static class PerformanceEndpoints
         var balances = accounts
             .Select(a => new PerfBalanceResponse(
                 a.Id.ToGuid(),
-                (ulong)a.CreditsPosted,
-                (ulong)a.DebitsPosted))
+                a.CreditsPosted,
+                a.DebitsPosted))
             .ToList();
 
         return Results.Ok(balances);
@@ -139,4 +139,4 @@ public record GetBalancesBatchRequest(List<Guid> AccountIds);
 
 public record CreateBatchResponse(int Count, List<Guid> Ids);
 
-public record PerfBalanceResponse(Guid AccountId, ulong CreditsPosted, ulong DebitsPosted);
+public record PerfBalanceResponse(Guid AccountId, UInt128 CreditsPosted, UInt128 DebitsPosted);
